fix(descriptions): unuse protable
